ScanAndSave/requirements.txt
python-dotenv
pillow
fastapi[standard]
pymysql
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.3.0
requests